fixed: undefined symbols in MythTV PVR client
The MythTV PVR client uses functionality from libmythxml.a but does not
link it in. Additionally, it fails to implement several PVR channel
functions, and has a wrong name for GetChannelsAmount(). Also, it is
linked against libdl which it doesn't use.

To fix the undefined symbols, make the client link against libmythxml.a,
add stubs for the missing functions, and change GetNumChannels() to
GetChannelsAmount(). Also drop unused -ldl from the linking command
line.